se hizo un componente del boton de salida de sala
