+++ src/components/room/lobby/components/ExitButton..vue
@@ -1,0 +1,21 @@
+<template>
+    <a v-if="socketStore.connected" @click="closeConnection" class="btn btn-error text-white">
+        <!-- boton salir sala -->
+        <Icon icon="ic:baseline-exit-to-app" width="24" height="24" />
+    </a>
+
+</template>
+
+<script setup>
+import { useWebSocketStore } from '../../stores/socketStore'
+import { Icon } from "@iconify/vue";
+
+const socketStore = useWebSocketStore();
+
+function closeConnection() {
+    socketStore.close();
+    window.location.href = '/protected/menu'
+};
+
+
+</script>
