+++ src/components/room/lobby/views/LobbyView.vue
@@ -44,16 +44,13 @@
       </v-card>
       <footer class="absolute bottom-0 z-10 flex flex-row justify-between py-2 px-4 w-full">
         <div>
-          <a v-if="socketStore.connected" @click="closeConnection" class="btn btn-error text-white">
-            <!-- boton salir sala -->
-            <Icon icon="ic:baseline-exit-to-app" width="24" height="24" />
-          </a>
-          <a v-else @click="connect" class="btn btn-warning text-white">
+          <ExitButton />
+          <a v-if="!socketStore.connected" @click="connect" class="btn btn-warning text-white">
             <Icon icon="ic:baseline-replay" width="24" height="24" />
           </a>
         </div>
         <div v-if="room.privileges" class="">
-          <QuorumButton buttonText="Iniciar" :action="startVoting"/>
+          <QuorumButton buttonText="Iniciar" :action="startVoting" />
         </div>
       </footer>
     </div>
@@ -79,11 +76,11 @@ import TabInfo from '../components/TabInfo.vue';
 import TabUsers from '../components/TabUsers.vue';
 import QuorumButton from '../components/QuorumButton.vue';
 import { Icon } from "@iconify/vue";
+import ExitButton from '../components/ExitButton..vue';
 
 const socketStore = useWebSocketStore();
 const {
   voting,
-  userList,
 } = storeToRefs(socketStore)
 
 
@@ -91,7 +88,7 @@ const router = useRouter()
 const room = ref('');
 const user = ref('');
 const quorum = ref('');
-let wsUrl = ''; 
+let wsUrl = '';
 
 const theme = ref(localStorage.getItem('theme') || 'mytheme');
 
@@ -121,19 +118,8 @@ onMounted(() => {
   socketStore.connect(wsUrl);
 });
 
-const isVotingDisabled = computed(() => {
-  // Bloquear si el quorum es mayor que la cantidad de usuarios conectados
-  return quorum.value > (userList.value ? userList.value.length : 0);
-});
-
-
 function connect() {
   socketStore.connect(wsUrl);
-};
-
-function closeConnection() {
-  socketStore.close();
-  window.location.href = '/protected/menu'
 };
 
 function startVoting() {
