+++ src/components/room/lobby/views/ResultsView.vue
@@ -1,5 +1,7 @@
 <template>
     <div class="p-10 flex flex-col gap-6 justify-center items-center">
+        <h1 v-if="socketStore.currentProposal.last_prop">Fin de votacion</h1>
+
         <h2>Resultados de la votación</h2>
         <div class="flex flex-row flex-wrap gap-4 justify-center ">
             <CardResult :result="displayResults" :proposal="socketStore.currentProposal" />
@@ -7,8 +9,11 @@
         </div>
 
         <VoteState />
-        <QuorumButton v-if="room.privileges && !socketStore.currentProposal.last_prop" buttonText="Siguiente propuesta" :action="nextProposal" />
-        <h1 v-if="socketStore.currentProposal.last_prop">fin de votacion</h1>
+        <QuorumButton v-if="room.privileges && !socketStore.currentProposal.last_prop" buttonText="Siguiente propuesta"
+            :action="nextProposal" />
+        <div v-if="socketStore.currentProposal.last_prop" class="flex flex-row justify-start w-full">
+            <ExitButton />
+        </div>
     </div>
 </template>
 
@@ -20,6 +25,7 @@ import VoteState from '../components/VoteState.vue';
 import { useRouter } from 'vue-router';
 import CardResult from '../components/CardResult.vue'
 import QuorumButton from '../components/QuorumButton.vue';
+import ExitButton from '../components/ExitButton..vue';
 
 const socketStore = useWebSocketStore();
